requirements.txt
asgiref==3.4.1
Django==3.2.8
pytz==2021.3
sqlparse==0.4.2
djangorestframework
feedparser
requests
python-dateutil==2.8.2